requirements.txt
# YOLO11 Security System - Requirements

# Core dependencies
ultralytics>=8.3.0
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# Data processing
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
albumentations>=1.3.0

# Web interface
streamlit>=1.28.0
plotly>=5.17.0
fastapi>=0.104.0
uvicorn>=0.24.0

# Utilities
pyyaml>=6.0
tqdm>=4.66.0
loguru>=0.7.0
python-dotenv>=1.0.0
schedule>=1.2.0
requests>=2.31.0

# Development
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.1.0
